Add JSX output option (#1)
* Add JSX output option.

This is done as a way of avoiding a dependency on `@babel/runtime`,
which forcibly pulls in polyfills even where they are not required.

The build is thus modified to perform minimal transpilation, with JSX
preserved. These modified sources can then be transpiled further as
needed by the consumer.

In addition, a dependency on `dom-helpers` is removed, because it
transitively added the `@babel/runtime` dependency as well.

* Update documentation for the fork.

* Remove `+automattic` from version number.

* Fix tests.

Fixed the import paths, and added steps to build UMD version of the
library for the tests, since we're not including it in the package.

* Add note about test changes to changelog.

* Small fix to pretest script.